requirement.txt
google-generativeai
python-dotenv
pytest